suggestions/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.5
watchdog==6.0.0
requests==2.32.5
python-levenshtein==0.27.3
